pyproject.toml
[project]
name = "python-automation"
version = "0.1.0"
description = ""
authors = [
    {name = "gunashree"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-behave (>=2.13.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
